# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Favorite Movies API",
    "version": "1.0.0",
    "description": "CRUD API for managing movie and review records in MongoDB Atlas."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://week3-week4-project.onrender.com",
      "description": "Render deployment URL"
    }
  ],
  "security": [
    { "sessionCookie": [] }
  ],
...
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "security": [],
        "tags": ["Authentication"],
        "summary": "Create a local account and sign in",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Credentials" },
              "example": { "email": "alex@example.com", "password": "SecurePassword123!" }
            }
          }
        },
        "responses": {
          "201": { "description": "Account created; session cookie set" },
          "400": { "description": "Invalid email or password length" },
          "409": { "description": "Email is already registered" },
          "429": { "description": "Too many authentication attempts" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "security": [],
        "tags": ["Authentication"],
        "summary": "Sign in with an email and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Credentials" },
              "example": { "email": "alex@example.com", "password": "SecurePassword123!" }
            }
          }
        },
        "responses": {
          "200": { "description": "Signed in; session cookie set" },
          "400": { "description": "Email and password are required" },
          "401": { "description": "Invalid email or password" },
          "429": { "description": "Too many authentication attempts" }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get the signed-in user's profile",
        "responses": {
          "200": { "description": "Current user profile" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Sign out and invalidate the current session",
        "responses": {
          "200": { "description": "Session invalidated and cookie cleared" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/auth/google": {
      "get": {
        "security": [],
        "tags": ["Authentication"],
        "summary": "Begin Google OAuth sign-in",
        "responses": {
          "302": { "description": "Redirect to Google consent" },
          "503": { "description": "Google OAuth credentials are not configured" }
        }
      }
    },
    "/api/auth/google/callback": {
      "get": {
        "security": [],
        "tags": ["Authentication"],
        "summary": "Complete Google OAuth sign-in",
        "responses": {
          "200": { "description": "Signed in; session cookie set" },
          "401": { "description": "Google account did not provide a verified email" },
          "503": { "description": "Google OAuth credentials are not configured" },
          "500": { "description": "OAuth callback could not be completed" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "movie.sid",
        "description": "HTTP-only session cookie issued by registration, login, or Google OAuth."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication is required; sign in and send the session cookie."
      }
    },
    "schemas": {
      "Credentials": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 72,
            "description": "12 to 72 UTF-8 bytes; stored as a bcrypt hash."
          }
        }
      },
...
    }
  }
}